Refactor revision worker to have its own base class.
This way, processor and revision workers are separated into their own
classes with shared functionality in the base class.
